leader changes direction
Added four anims to Kris for up, down, left, and right movement.
Added two functions to pcharDisp (the party character tmeplate) to stop the current animation, or play an animation based on the sent name.

In pcharDisp, added a check in move_step that checks move_dir to change the animation.